% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embark_palette.R
\docType{data}
\name{embark_colors}
\alias{embark_colors}
\title{Embark Lab Color Palette}
\format{
A named character vector with 6 hex color values.
}
\usage{
embark_colors
}
\description{
A named character vector of six Embark Lab brand colors.
}
\keyword{datasets}
