% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embark_palette.R
\name{embark_palette_named}
\alias{embark_palette_named}
\title{Get Named Embark Color Palette}
\usage{
embark_palette_named()
}
\value{
A named character vector of hex color values.
}
\description{
Returns the Embark Lab color palette with names intact.
}
