% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embark_themes.R
\docType{data}
\name{embark_theme_a}
\alias{embark_theme_a}
\title{Embark ggplot2 Theme (Avenir Font)}
\format{
A ggplot2 theme object.
}
\usage{
embark_theme_a
}
\description{
A ggplot2 theme using the Avenir font family with Embark Lab Navy branding.
Based on \code{\link[ggplot2:theme_minimal]{ggplot2::theme_minimal()}} with centered bold titles, top legend,
and transparent background.
}
\keyword{datasets}
