% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embark_palette.R
\name{scale_fill_embark}
\alias{scale_fill_embark}
\title{Embark Fill Scale for ggplot2}
\usage{
scale_fill_embark(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}}.}
}
\value{
A ggplot2 scale object.
}
\description{
A convenience wrapper around \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}} using the
Embark Lab color palette.
}
