% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embark_palette.R
\docType{data}
\name{embark_colors}
\alias{embark_colors}
\title{embark_colors
adds embark color palette}
\format{
An object of class \code{character} of length 6.
}
\usage{
embark_colors
}
\value{
color palette
}
\description{
embark_colors
adds embark color palette
}
\keyword{datasets}
