% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embark_palette.R
\name{embark_palette}
\alias{embark_palette}
\title{embark_palette
adds embark color palette}
\usage{
embark_palette(n = NULL, names = FALSE)
}
\arguments{
\item{n}{number of colors}

\item{return_names}{return color names, T or F}
}
\value{
color palette
}
\description{
embark_palette
adds embark color palette
}
