% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embark_palette.R
\name{embark_palette_named}
\alias{embark_palette_named}
\title{embark_palette_named
adds embark color palette with names}
\usage{
embark_palette_named()
}
\value{
color palette
}
\description{
embark_palette_named
adds embark color palette with names
}
