% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embark_themes.R
\docType{data}
\name{embark_theme_a}
\alias{embark_theme_a}
\title{embark_theme_a
adds theme with Avenir font family}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 97.
}
\usage{
embark_theme_a
}
\value{
ggtheme
}
\description{
embark_theme_a
adds theme with Avenir font family
}
\keyword{datasets}
