% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embark_themes.R
\docType{data}
\name{embark_theme_g}
\alias{embark_theme_g}
\title{embark_theme_g
adds theme with Garet font family}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 97.
}
\usage{
embark_theme_g
}
\value{
ggtheme
}
\description{
embark_theme_g
adds theme with Garet font family
}
\keyword{datasets}
