% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_by_age_sex.R
\name{frq_table_by_age_sex}
\alias{frq_table_by_age_sex}
\title{Frequency Table by Age and Sex Function
creates a table which summarizes responses on an identified variable across age and sex}
\usage{
frq_table_by_age_sex(df, x_var)
}
\arguments{
\item{df}{A dataframe/tibble}

\item{x_var}{A variable}
}
\value{
a table which summarizes responses on an identified variable across age and sex
}
\description{
Frequency Table by Age and Sex Function
creates a table which summarizes responses on an identified variable across age and sex
}
