% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_by_age_sex.R
\name{hist_by_age_sex}
\alias{hist_by_age_sex}
\title{hist_by_age_sex function
creates a histogram (by percent) an identified variable across age and sex}
\usage{
hist_by_age_sex(df_1, x_var_1, graph_title)
}
\arguments{
\item{df_1}{a dataframe}

\item{x_var_1}{a variable for histogram}

\item{graph_title}{title of the graph}
}
\value{
a histogram (percentages within gender) of an identified variable across age
}
\description{
hist_by_age_sex function
creates a histogram (by percent) an identified variable across age and sex
}
