% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embark_palette.R
\name{scale_color_embark}
\alias{scale_color_embark}
\title{scale_color_embark
colors graph with embark palette}
\usage{
scale_color_embark(...)
}
\description{
scale_color_embark
colors graph with embark palette
}
