% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embark_palette.R
\name{scale_fill_embark}
\alias{scale_fill_embark}
\title{scale_fill_embark
fills graph with embark palette}
\usage{
scale_fill_embark(...)
}
\description{
scale_fill_embark
fills graph with embark palette
}
